Move `ResolveTypeNames` to outputable
Since magic comments are per-file, it makes more sense to handle them with `CreateFileByName`.
