--- sig/orthoses/resolve_type_names.rbs
@@ -1,19 +1,0 @@
-# THIS IS GENERATED CODE from `$ rake sig`
-
-# use Orthoses::ResolveTypeNames
-class Orthoses::ResolveTypeNames
-  @loader: untyped
-  def initialize: (untyped loader) -> void
-  def call: () -> untyped
-  private def entry_and_members: (untyped env, untyped type_name) -> untyped
-  private def content_header: (untyped entry) -> untyped
-  private def class_header: (untyped decl) -> ::String
-  private def module_header: (untyped decl) -> ::String
-  private def interface_header: (untyped decl) -> ::String
-end
-
-module Orthoses::ResolveTypeNames::WriterCopy
-  def name_and_args: (untyped name, untyped args) -> (::String | nil)
-
-  def name_and_params: (untyped name, untyped params) -> ::String
-end
